move some stuff out of simpleEvaluate
